Conflicts:
	gameplay-encoder/gameplay-encoder.vcxproj.filters
	gameplay-encoder/src/AnimationChannel.cpp
	gameplay-encoder/src/AnimationChannel.h
	gameplay-encoder/src/DAEChannelTarget.cpp
	gameplay-encoder/src/DAESceneEncoder.cpp
	gameplay-encoder/src/EncoderArguments.cpp
	gameplay-encoder/src/FileIO.cpp
	gameplay-encoder/src/FileIO.h
	gameplay-encoder/src/GPBFile.cpp
	gameplay-encoder/src/Mesh.cpp
	gameplay-encoder/src/MeshSkin.cpp
	gameplay-encoder/src/MeshSkin.h
	gameplay-encoder/src/Node.cpp
	gameplay-encoder/src/Node.h
	gameplay-encoder/src/Scene.h
	gameplay-encoder/src/StringUtil.cpp
	gameplay-encoder/src/StringUtil.h
	gameplay-encoder/src/Vector2.cpp
	gameplay-encoder/src/Vector2.h
	gameplay-encoder/src/Vector3.h
	gameplay-encoder/src/Vector4.h
	gameplay-encoder/src/Vertex.cpp
	gameplay/src/Joint.cpp

+++ gameplay/src/Ray.cpp
@@ -135,7 +135,7 @@ void Ray::normalize()
         return;
 
     // Normalize the ray's direction vector.
-    float normalizeFactor = 1.0f / sqrtf(_direction.x * _direction.x + _direction.y * _direction.y + _direction.z * _direction.z);
+    float normalizeFactor = 1.0f / sqrt(_direction.x * _direction.x + _direction.y * _direction.y + _direction.z * _direction.z);
 
     if ( normalizeFactor != 1.0f )
     {
